+++ console/src/consoleUtils.h
@@ -18,7 +18,7 @@ typedef enum
 
 typedef struct
 {
-	int size;
+	uint32_t size;
 	void* stream;
 } t_buffer;
 
@@ -30,13 +30,13 @@ typedef struct
 
 
 
-int crear_conexion(t_log*, const char* server_name, char *ip, char* puerto);
-void enviar_mensaje(char* mensaje, int socket_cliente);
+uint32_t crear_conexion(t_log*, const char* server_name, char *ip, char* puerto);
+void enviar_mensaje(char* mensaje, uint32_t socket_cliente);
 t_paquete* crear_paquete(void);
 t_paquete* crear_super_paquete(void);
-void agregar_a_paquete(t_paquete* paquete, void* valor, int tamanio);
-void enviar_paquete(t_paquete* paquete, int8_t socket_cliente);
-void liberar_conexion(int socket_cliente);
+void agregar_a_paquete(t_paquete* paquete, void* valor, uint32_t tamanio);
+void enviar_paquete(t_paquete* paquete, uint32_t socket_cliente);
+void liberar_conexion(uint32_t socket_cliente);
 void eliminar_paquete(t_paquete* paquete);
 
 #endif /* UTILS_H_ */
